Smart_camera_recog_demo_backend/requirements.txt
asgiref==3.7.2
awscrt==0.16.23
AWSIoTPythonSDK==1.5.2
awsiotsdk==1.16.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-cors-headers==4.1.0
idna==3.4
numpy==1.24.4
opencv-python==4.8.0.74
paho-mqtt==1.6.1
Pillow==10.0.0
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.6.3
urllib3==2.0.3
